requirements.txt
# Framework MCP
fastmcp>=0.1.0

huggingface_hub

# API et serveurs
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

langchain-groq>=0.2.0
langchain_community
langchain-mistralai

# Frontend et interface (optionnel)
streamlit>=1.37.0

# LangChain et LangGraph
langchain>=0.3.0
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-google-genai>=1.0.0

# LLM clients
openai>=1.50.0
google-generativeai>=0.8.0

# Bases de données
psycopg2-binary>=2.9.0

# Utilitaires
python-dotenv>=1.0.0
nest-asyncio>=1.6.0
pandas>=2.2.0
pydantic>=2.0.0
anyio>=4.0.0